Add unit tests for GET request, fixes AB#348
